Moves existing exceptions to exception module.
