--- ocean_lib/common/aquarius/exceptions.py
@@ -1,8 +1,0 @@
-#
-# Copyright 2021 Ocean Protocol Foundation
-# SPDX-License-Identifier: Apache-2.0
-#
-
-
-class AquariusGenericError(Exception):
-    """ Error invoking an Aquarius metadata service endpoint"""
