+++ ocean_lib/data_provider/data_service_provider.py
@@ -13,11 +13,11 @@ from json import JSONDecodeError
 
 import requests
 from ocean_lib.common.agreements.service_types import ServiceTypes
-from ocean_lib.common.exceptions import OceanEncryptAssetUrlsError
 from ocean_lib.common.http_requests.requests_session import get_requests_session
 from ocean_lib.config_provider import ConfigProvider
 from ocean_lib.data_provider.exceptions import InvalidURLException
 from ocean_lib.enforce_typing_shim import enforce_types_shim
+from ocean_lib.exceptions import OceanEncryptAssetUrlsError
 from ocean_lib.models.algorithm_metadata import AlgorithmMetadata
 from ocean_lib.ocean.env_constants import ENV_PROVIDER_API_VERSION
 from ocean_lib.web3_internal.utils import add_ethereum_prefix_and_hash_msg
