+++ ocean_lib/data_provider/test/test_data_service_provider.py
@@ -6,12 +6,12 @@
 from unittest.mock import Mock
 
 import pytest
-from ocean_lib.common.exceptions import OceanEncryptAssetUrlsError
 from ocean_lib.common.http_requests.requests_session import get_requests_session
 from ocean_lib.config_provider import ConfigProvider
 from ocean_lib.data_provider.data_service_provider import DataServiceProvider as DataSP
 from ocean_lib.data_provider.data_service_provider import urljoin
 from ocean_lib.data_provider.exceptions import InvalidURLException
+from ocean_lib.exceptions import OceanEncryptAssetUrlsError
 from requests.models import Response
 from tests.resources.helper_functions import get_publisher_ocean_instance
 from tests.resources.mocks.http_client_mock import (
